telegram-bot-service/requirements.txt
aiogram>=3.13.0
motor>=3.6.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.10.0,<3.13
cryptography>=41.0.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0

